Consolidate duplicate code and replace inline reimplementations with shared utilities
- Extract identical GitHub client mock (GitHubAuthError, GitHubRateLimitError,
  GitHubPermanentError + ghGet stub) that was copy-pasted across two
  PRPollingSource test files into src/test-kernel/support/githubClientMock.ts;
  both test files now call mockGitHubClient(ghGet) instead.
- Remove local nonEmptyString helper in workPlan.ts; replace with
  isNonEmptyString from @utils/core (same semantics, already available).
- Replace inline typeof/trim check in modelSwitchState.ts with isNonEmptyString
  from @utils/core to match the codebase-wide convention.
